Problem in ant.c probabily select

+++ tsp_solver.h
@@ -11,6 +11,6 @@ struct _tsp_solver{
 
 struct _tsp_solver* init_tsp_solver(struct _graph *graph,int numofants);
 void freeing_tsp_solver(struct _tsp_solver*);
-void solve_tsp(struct _graph*,struct _tsp_solver*,int);
+void solve_tsp(struct _graph*,struct _tsp_solver*,double);
 
 #endif
